Move JulianDateNumber to DateUtil
